+++ milestone-3/frontend/src/components/Home.tsx
@@ -1,7 +1,15 @@
 import React from "react";
 import "./home.css";
 import { recipes } from "../recipeData";
+import RecipePreview from "./RecipePreview";
 
+export interface Recipe {
+    name: string;
+    description: string;
+    image: string;
+    ingredients: string[];
+    instructions: string[];
+  }
 
 export default function Home() {
   return (
@@ -11,8 +19,10 @@ export default function Home() {
         <div className="container">
             <h1>Wecome To My Kitchen</h1>
             <p>I wish everyone who loves to cook would have this website bookmarked:)</p>
-            {recipes.map(recipe => 
-                <h1>test</h1>
+            {recipes.map((recipe?:any) => 
+                <RecipePreview 
+                    {...recipe}
+                />
             )}
             
         </div>
