+++ milestone-3/frontend/src/recipeData.ts
@@ -1,17 +1,20 @@
-export interface Recipe {
-    name: string;
-    description: string;
-    image: string;
-    ingredients: string[];
-    instructions: string[];
-  }
+// export interface Recipe {
+//     name: string;
+//     description: string;
+//     image: string;
+//     ingredients: string[];
+//     instructions: string[];
+//   }
+
+import sushiImg from './image/sushi.jpg'
+
 
 
 export const recipes=[
     {
         name: 'Spicy Crab Roll',
         description: 'Spicy Crab Roll is a popular Japanese sushi roll – creamy, satisfying, and fiery spicy.',
-        image: './image/sushi.jpg',
+        image: sushiImg,
         ingredients: ["The Rice", "1 Teaspoon of The Spice", "The Sushi Mat" ],
         instructions: ["Getting the rice right is the trickiest part of making sushi. If you end up with enough rice for an extra roll, make an experimental one using whatever you’ve got in your fridge: leftover meat, herbs, etc.",
                        "For ground cayenne, we recommend starting with 1/4 teaspoon, tasting the mixture, and adding more if you like.",
@@ -21,7 +24,7 @@ export const recipes=[
     {
         name: '2',
         description: 'Spicy Crab Roll is a popular Japanese sushi roll – creamy, satisfying, and fiery spicy.',
-        image: './image/sushi.jpg',
+        image:sushiImg,
         ingredients: ["The Rice", "1 Teaspoon of The Spice", "The Sushi Mat" ],
         instructions: ["Getting the rice right is the trickiest part of making sushi. If you end up with enough rice for an extra roll, make an experimental one using whatever you’ve got in your fridge: leftover meat, herbs, etc.",
                        "For ground cayenne, we recommend starting with 1/4 teaspoon, tasting the mixture, and adding more if you like.",
@@ -31,7 +34,7 @@ export const recipes=[
     {
         name: '3',
         description: 'Spicy Crab Roll is a popular Japanese sushi roll – creamy, satisfying, and fiery spicy.',
-        image: './image/sushi.jpg',
+        image: sushiImg,
         ingredients: ["The Rice", "1 Teaspoon of The Spice", "The Sushi Mat" ],
         instructions: ["Getting the rice right is the trickiest part of making sushi. If you end up with enough rice for an extra roll, make an experimental one using whatever you’ve got in your fridge: leftover meat, herbs, etc.",
                        "For ground cayenne, we recommend starting with 1/4 teaspoon, tasting the mixture, and adding more if you like.",
